fix C4267 warning on Windows and re-enable C4267 warning for windows build
Reviewed-on: https://git-ccr-1.devtools.intel.com/gerrit/53598
Tested-by: webrtctest <webrtctest@intel.com>

+++ talk/woogeen/sdk/base/customizedaudiocapturer.cc
@@ -409,8 +409,9 @@ bool CustomizedAudioCapturer::RecThreadProcess() {
 
   if (last_call_record_millis_ == 0 ||
       current_time - last_call_record_millis_ >= 10) {
-    if (frame_generator_->GenerateFramesForNext10Ms(recording_buffer_.get(),
-                                                    recording_buffer_size_) !=
+    if (frame_generator_->GenerateFramesForNext10Ms(
+            recording_buffer_.get(),
+            static_cast<uint32_t>(recording_buffer_size_)) !=
         static_cast<uint32_t>(recording_buffer_size_)) {
       crit_sect_.Leave();
       RTC_DCHECK(false);
